+++ go.mod
@@ -3,13 +3,12 @@ module github.com/dallasurbanists/events-sync
 go 1.24.4
 
 require (
+	github.com/golang-migrate/migrate/v4 v4.18.3
 	github.com/jmoiron/sqlx v1.4.0
 	github.com/lib/pq v1.10.9
 )
 
 require (
-	github.com/golang-migrate/migrate/v4 v4.18.3 // indirect
-	github.com/gorilla/mux v1.8.1 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
