+++ internal/server/router.go
@@ -1,0 +1,15 @@
+package server
+
+import "net/http"
+
+func (s *Server) newConfiguredRouter() *http.ServeMux {
+	router := http.NewServeMux()
+
+	router.Handle("GET /", http.FileServer(http.Dir("web")))
+
+	router.HandleFunc("GET /api/events", s.getEvents)
+	router.HandleFunc("PUT /api/events/{uid}/status", s.updateEventStatus)
+	router.HandleFunc("GET /api/events/stats", s.getEventStats)
+
+	return router
+}
