--- internal/server/server_test.go
@@ -1,15 +1,0 @@
-package server
-
-import (
-	"testing"
-)
-
-func TestNewServer(t *testing.T) {
-	// This is a basic test to ensure the server can be created
-	// In a real application, you'd want to mock the database
-	// and test the HTTP endpoints
-
-	// For now, we'll just test that the package compiles
-	// and the NewServer function exists
-	_ = NewServer
-}
